Update README.md
We created a product that assisted digital nomads in making informed decisions about their workations, with a specific focus on European countries. To shape our product, we primarily utilized artifacts such as Python, PostgreSQL, Tableau, and a Microsoft Azure cloud-based database. Working as a team of four, we employed agile methods to quickly sculpt our project according to our requirements.